refactor(ui): rename SnackbarNew to Snackbar
