+++ ui/src/layouts/SimpleLayout.vue
@@ -1,6 +1,6 @@
 <template>
   <v-app>
-    <SnackbarNew />
+    <Snackbar />
     <v-main class="d-flex justify-center align-center">
       <v-container
         class="full-height d-flex justify-center align-center"
@@ -41,7 +41,7 @@
 import { computed } from "vue";
 import { useRouter } from "vue-router";
 import Logo from "../assets/logo-inverted.png";
-import SnackbarNew from "@/components/Snackbar/SnackbarNew.vue";
+import Snackbar from "@/components/Snackbar/Snackbar.vue";
 
 const router = useRouter();
 const currentRoute = computed(() => router.currentRoute);
